- resolved warnings about possible loss of data by using static_cast

+++ src/backends/mysql/standard-use-type.cpp
@@ -65,7 +65,7 @@ void mysql_standard_use_type_backend::pre_use(indicator const *ind)
             {
                 std::string const& s = exchange_type_cast<x_stdstring>(data_);
                 buf_ = quote(statement_.session_.conn_,
-                             s.c_str(), s.size());
+                             s.c_str(), static_cast<int>(s.size()));
             }
             break;
         case x_short:
